pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "firm-protocol"
version = "1.1.0"
description = "FIRM Protocol — Self-Evolving Autonomous Organization Runtime"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Romain Santoli"},
    {name = "FIRM Protocol Contributors"},
]
keywords = [
    "autonomous-organizations",
    "multi-agent",
    "self-evolving",
    "governance",
    "protocol",
    "llm",
    "ai-agents",
    "firm",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]
# Core has zero external dependencies — stdlib only
dependencies = []

[project.scripts]
firm = "firm.cli:main"

[project.optional-dependencies]
llm = [
    "openai>=1.30",
    "anthropic>=0.25",
    "mistralai>=1.0",
    "httpx>=0.27",
]
bounty = [
    "httpx>=0.27",
    "pyyaml>=6.0",
]
api = [
    "fastapi>=0.111",
    "pydantic>=2.0",
    "uvicorn>=0.30",
]
all = [
    "firm-protocol[llm,api,bounty]",
]
dev = [
    "firm-protocol[all]",
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "hypothesis>=6.0",
]
docs = [
    "sphinx>=7.0",
    "sphinx-autodoc-typehints>=2.0",
]

[project.urls]
Homepage = "https://github.com/romainsantoli-web/firm-protocol"
Repository = "https://github.com/romainsantoli-web/firm-protocol"
Issues = "https://github.com/romainsantoli-web/firm-protocol/issues"
Changelog = "https://github.com/romainsantoli-web/firm-protocol/blob/main/CHANGELOG.md"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --cov=src/firm --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
source = ["src/firm"]
omit = [
    "src/firm/llm/*",
    "src/firm/api/*",
    "src/firm/__main__.py",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.setuptools.packages.find]
where = ["src"]
